+++ changelog.md
@@ -9,6 +9,7 @@
 * **[Modding]** Updated support for F/A-18E/F/G mod version 2.2.5
 * **[Campaign Setup]** Allow adjustments to naval TGOs (except carriers) on turn 0
 * **[Campaign Design]** Ability to configure specific carrier names & types in campaign's yaml file 
+* **[Mission Generation]** Ability to inject custom kneeboards
 
 ## Fixes
 * **[UI/UX]** A-10A flights can be edited again.